github/requirements.txt
streamlit
langchain
sentence-transformers
pinecone-client
PyPDF2
pydantic
langgraph
os
langchain_groq